convert to mutable_data_ptr data_ptr calls immediately after at::empty() (#98734)
The tensor is uninitialized in this case, so it is highly likely to be
written before it will be read. Furthermore, because it is a new
tensor, there's no harm in getting mutable access to it: there's no
lazy copy that would be materialized.

This was automatically generated with a regular expression.

Differential Revision: [D44831830](https://our.internmc.facebook.com/intern/diff/D44831830/)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/98734
Approved by: https://github.com/ezyang