Add torch_lazy_enable_device_data_cache to disable lazy device data cache (#107827)
### Add python binding variables for enabling and disabling

These changes will be used in the pytorch/xla repository for lowering HLO for the AWS Neuron compiler.  For correct tensor lowerings the device cache size must be set to zero.

It is advantageous to be able to enable and disable the cache without deleting it.  This allows use of the XLA device, and HLO lowering in a single python file, isolating cache disablement to a python context.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/107827
Approved by: https://github.com/JackCaoG, https://github.com/wconstab, https://github.com/bdhirsh