[2D][2/N][DeviceMesh] Add get_parent_mesh_dim() in _MeshEnv class (#109330)
Adding some additional APIs that are needed for 2D workflow.

Since each parallelism is only aware of its own mesh when we are constructing 2D state_dict. We need to know the mesh_dim of the child mesh in the parent mesh. So, we can use it to create DTensor that is 2D sound.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109330
Approved by: https://github.com/fegin, https://github.com/fduwjj, https://github.com/wanchaol