[c10d] Fix the hang issue in store.check(TIMEOUT_DUMP) (#116297)
Summary:
We have found out the root cause of the hang is NOT due to destruction
of stores. The hang in the check() only happens when the store is of
type FileStore.

The file held by each filestore was a temp file, which was created by
Python Tempfile, it was deleted by default when the file was closed.

Note that the file was opened and closed by every check() in the watchdog and in constructor of FileStore.

The when check() tried to open the deleted file again, open() would fail
after the timeout value (by default 5 mins), hence the hang happened.

The fix is simple, just avoid the default deletion after the file is
closed.
Test Plan:

1. We first reproduce the hang in check() in the existing unit test:
   test_init_process_group_for_all_backends by enabling the
   DumpOnTimeOut and making the main thread sleep for 2s, to give enough time for tempfile
   to be deleted
2. Adding log to check ref count of fileStore and also the sequence of
   file opening and closing
3. With the repro, an exception will be thrown as "no such file or
   directory' and unit test would fail
4. Verify the tests now passes with the above knob change
5. add an unit test in test_c10d_nccl to cover the fileStore check() code path
python test/distributed/test_c10d_common.py ProcessGroupWithDispatchedCollectivesTests
python test/distributed/test_c10d_nccl.py ProcessGroupNCCLTest.test_file_store_check
Reviewers:

Subscribers:

Tasks:
T173200093
Tags:

Pull Request resolved: https://github.com/pytorch/pytorch/pull/116297
Approved by: https://github.com/fduwjj
ghstack dependencies: #116296