Update on "torch.utils.checkpoint preserves torch function mode stack during recompute"
Fixes https://github.com/pytorch/pytorch/issues/147995

TorchFunctionModeTLS is part of the autograd tls, but because .backward() itself is a leaf for TorchFunctionMode, the mode is disabled before we enter into the engine.  Conversely, since TorchDispatchMode traces through the .backward() python call, we don't actually need to manually stash/restore if the user keeps the same mode enabled.

We should still fix TorchDispatchMode though, because even if the user doesn't keep the same mode enabled on the .backward() call, checkpoint should not fail.

[ghstack-poisoned]