Introduce stages to aot_dispatch
The starting point for this refactor is that I need access to the fully
general joint graph representation in an export-like interface, but I
then subsequently need a way to feed this joint graph into the rest of
the compilation pipeline so I can get an actual callable that I can run
once I've finished modifying it.  Previously, people had added export
capabilities to AOTAutograd by having an export flag that toggled what
exactly the functions return and triggering aot_dispatch to go to a
different "export" implementation, but I've found this difficult to
understand and has lead to a bit of duplicate code for the export path.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
ghstack-source-id: 98401e07b420cc918f7e7c807bd6b0c535b7b451
Pull-Request: https://github.com/pytorch/pytorch/pull/158213