Do not decompose in functionalization/proxy tensor if autograd wouldn't have decomposed
This fixes AOTAutograd rms_norm not being bitwise equivalent to
eager, because it avoids a decomposition.  You can force the
decomposition by having the decomposition in the dispatch table,
but if eager mode wouldn't have decomposed (because it went to the fused
one), we now default to preserving the fused call by default.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
ghstack-source-id: d0a7f2837ad9eb2c0209a7404d21e59abebf38f4
Pull-Request: https://github.com/pytorch/pytorch/pull/164939