[dynamo][ac] Config flag to allow eager and compile AC divergence for side-effects
Eager AC/SAC reapplies the mutations (like global dict mutations) in the backward during the recomputation of forward. torch.compile has no easy way to reapply python mutations in the backward. But many users might be ok to skip reapplication of side effects in the backward. They can set this config flag to accept this eager and compile divergence.

ghstack-source-id: 148e19647deea81513e556103d2427ba54fecf57
Pull Request resolved: https://github.com/pytorch/pytorch/pull/165775