Add some code for exploring the space of accessible size/stride configs via plain views
We are working on a translation from as_strided to view operations, but
only when the as_strided is representable as a plain view.  A useful
testing utility in this situation is the ability to enumerate all valid
views on an original tensor.  So we have a small test here that shows
it is possible.

To avoid an explosion of states, we don't handle permutes and size=1,
which are degenerate cases (you can always do a single permute and
a series of unsqueezes to get to the final desired state.)

Authored with claude code assistance.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
ghstack-source-id: b70e19140387b066059bc2ab2ca44db2902c0ca4
Pull-Request: https://github.com/pytorch/pytorch/pull/167076